requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pdf2image==1.17.0
pillow==10.4.0
PyPDF2==3.0.1
typing_extensions==4.12.2
Werkzeug==3.0.4
zipp==3.20.2
